requirements.txt
Django==4.2.7

gunicorn==21.2.0
uvicorn==0.23.2

# for sentry error monitoring
sentry-sdk==1.29.2

# for creating REST APIs
djangorestframework==3.14.0

# for swagger doc
drf-yasg == 1.21.7

# to support json field
django-jsonfield==1.4.1

# token
djangorestframework-simplejwt==5.2.0

# for CORS
django-cors-headers==4.1.0

# for testing
pytest==7.4.0
pytest-django==4.5.2
factory_boy==3.2.1
pytest-cov==4.1.0


# for code formatting
black==22.8.0

# mysql python support for Python
mysqlclient

# postgresql support for Python
psycopg2-binary==2.9.7


# for getting env
django-environ==0.10.0

# python linter
ruff==0.0.281

# for download file as excel
openpyxl==3.1.2

# for generating pdf
weasyprint==60.1


# as name defined - for git pre-commit
pre-commit==3.3.3


redis==4.4.4

# Freeze time for tests
freezegun==1.2.2

# celery for running jobs
celery==5.3.4
# run task on schedule
django-celery-beat==2.5.0

# send email via celery
django-celery-email==3.0.0

djangorestframework-api-key==3.0.0

# for serving static files
whitenoise==6.6.0

# Bypasses ALLOWED_HOSTS for the purposes of load balancers
django-lb-health-check==1.0.1
psycopg2-binary==2.9.9
requests